requirements.txt
-e git+https://github.com/django/django.git@9a2bceed1aab52f65820c378f5ae1f608322b55c#egg=Django-dev
South==0.7.6
Whoosh==2.4.1
argparse==1.2.1
dj-database-url==0.2.1
django-grappelli==2.4.2
-e git+https://github.com/toastdriven/django-haystack.git@13e5a4926992581106ed14db1f8dd58a897c9a76#egg=django_haystack-dev
psycopg2==2.4.5
python-ptrace==0.6.4
wsgiref==0.1.2
